[Chore/#174] OFRAlertAction 별도 파일로 분리
